Visitor pattern

+++ java/src/ball/BallPosition.java
@@ -3,27 +3,27 @@ package ball;
 import com.google.gson.Gson;
 
 public class BallPosition {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public BallPosition(int x, int y) {
+    public BallPosition(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
